gem-location: support remote file
When you are working on a remote directory with tramp mode, bundle-open
failed to open specified gem directory.  Check the default-directory for
a remote identifier associated to the current buffer and prepend to the
gem-location string.

If the current directory is local, the default-directory is nil and
concat silently convert it to an empty string, which makes this change
effectively a no-op.